Changed Timer to System.nanoTime()
There are many options for timers in android all with their own
problems.

Ones Considered:
- (Previously being used) SystemClock.uptimeMillis() has problems with
sleep and can drift.
- SystemClock.elapsedTime() has problems with drifting.
- System.currentTimeMillis() has well known issues with clock changes.
- System.nanoTime() based off boot time and very stable with
drift/sleep.

So with that System.nanoTime() became the new basis for the timers used.

* New Timer based on System.nanoTime()
* Removed unnecessary complication introduced with MissionTimer.java
and merged couple methods used down to StopWatch.java.
* Moved most timing to use nano seconds to be aligned with new timer.

+++ app/src/main/java/com/boarbeard/audio/parser/GermanParser.java
@@ -1,12 +1,12 @@
 package com.boarbeard.audio.parser;
 
-import java.util.Collections;
-import java.util.EnumSet;
-import java.util.Set;
-
 import com.boarbeard.audio.MediaPlayerSequence;
 import com.boarbeard.audio.parser.Grammar.Element;
 import com.boarbeard.generator.beimax.event.Threat;
+
+import java.util.Collections;
+import java.util.EnumSet;
+import java.util.Set;
 
 /**
  * Parser which supports German Grammars.
@@ -29,7 +29,7 @@ public class GermanParser extends EnglishParser {
 
 	@Override
 	protected void appendThreatLocation(StringBuilder sb, Threat event,
-			int startTime, MediaPlayerSequence output) {
+			long startTime, MediaPlayerSequence output) {
 		final Element location;
 
 		if (event.getThreatLevel() == Threat.THREAT_LEVEL_SERIOUS) {
